New dotfiles
The bare git repo idea is neato, but a bit of a hassle to setup quickly
(and when I'm needing to do it, playing with this setup isn't what I'm
wanting to focus on).

Also made some changes to the configs themselves and replaced the vim
stuff with an init script.

Additionally, remove the awesome config stuff.